// src/app/models/registration.model.ts
export class Registration {
  id?: any;
  firstname?: string;
  lastname?: string;
  email?: string;
  phone?: string;
  age?: number;
  city?: string;
  gender?: string;
  arrivals?: string;
  productinfo?: string;
  residential?: string;
  type?: string;
  amount?: string;
  txnid?: string;
  paym?: string;
  status?: string;
  agree?: boolean;
}